Started hub/listener

+++ src/main/java/org/sims/controller/HubController.java
@@ -4,11 +4,14 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ser.impl.SimpleBeanPropertyFilter;
 import com.fasterxml.jackson.databind.ser.impl.SimpleFilterProvider;
-import org.sims.model.Hub;
-import org.sims.model.SpecificNotification;
+import com.querydsl.core.BooleanBuilder;
+import com.querydsl.core.types.Predicate;
+import org.sims.model.*;
 import org.sims.repository.HubRepository;
-import org.sims.repository.SpecificNotificationRepository;
+//import org.sims.repository.SpecificNotificationRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.rest.webmvc.ResourceNotFoundException;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.converter.json.MappingJacksonValue;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
@@ -19,66 +22,43 @@ import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 
-@ApiIgnore
 @RestController
 @CrossOrigin("*")
 @RequestMapping("/api")
 public class HubController {
 
   private HubRepository hubRepository;
-  private SpecificNotificationRepository specificNotificationRepository;
+//  private SpecificNotificationRepository specificNotificationRepository;
 
   @Autowired
-  public HubController(HubRepository hubRepository, SpecificNotificationRepository specificNotificationRepository) {
+  public HubController(HubRepository hubRepository/*, SpecificNotificationRepository specificNotificationRepository*/) {
     this.hubRepository = hubRepository;
-    this.specificNotificationRepository = specificNotificationRepository;
+//    this.specificNotificationRepository = specificNotificationRepository;
   }
 
+  @ResponseStatus(HttpStatus.CREATED)
   @PostMapping("/hub")
-  public Hub registerListener(@RequestBody String requestBody) {
-
-    MultiValueMap<String, String> callback = new LinkedMultiValueMap<>();
-    if(requestBody != null) {
-      callback.add("callback", requestBody);
-    }
-
-    System.out.println(requestBody);
-    System.out.println(callback);
-
-    ObjectMapper mapper = new ObjectMapper();
-
-    TypeReference<HashMap<String, String>> typeRef
-            = new TypeReference<HashMap<String, String>>() {};
-    Map<String, String> map;
-    try {
-      map = mapper.readValue(requestBody, typeRef);
-      System.out.println(map);
-      System.out.println(map.get("callback"));
-      Hub hub = new Hub();
-      hub.setCallback(map.get("callback"));
-      return hubRepository.save(hub);
-
-    } catch (IOException e) {
-      e.printStackTrace();
-    }
-    return null;
-
-
+  public Hub registerListener(@RequestBody Hub hub) {
+    return hubRepository.save(hub);
   }
 
   @GetMapping("/hub")
-  public MappingJacksonValue getHub() {
-    MappingJacksonValue mappingJacksonValue = new MappingJacksonValue(specificNotificationRepository.findAll());
-    SimpleFilterProvider filters;
-    filters = (new SimpleFilterProvider()).addFilter("service",
-            SimpleBeanPropertyFilter.serializeAll());
-    mappingJacksonValue.setFilters(filters);
-    return mappingJacksonValue;
+  public List<Hub> getHub() {
+    return hubRepository.findAll();
   }
 
+  @ResponseStatus(HttpStatus.NO_CONTENT)
   @DeleteMapping("/hub/{id}")
   public void unregisterListener(@PathVariable String id) {
+    QHub qHub = QHub.hub;
+    Predicate p = new BooleanBuilder();
+    ((BooleanBuilder) p).and(qHub.id.eq(id));
+    Optional<Hub> optionalHub = hubRepository.findOne(p);
+    if(!optionalHub.isPresent()) {
+      throw new ResourceNotFoundException();
+    }
     hubRepository.deleteById(id);
   }
